requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0

# LangChain Ecosystem (Known Good Set)
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.5
langchain-openai==0.1.8
langgraph==0.1.5

# OpenAI
openai==1.35.7